requirements.txt
django==5.0.7
djangorestframework==3.15.2
psycopg2-binary==2.9.9
djangorestframework-simplejwt==5.3.1
celery==5.3.6
redis==5.0.1
flower==2.0.1
drf-yasg==1.21.7
django-phonenumbers==1.0.1
django-phonenumber-field==8.0.0
django-filter==24.2
django-celery-beat==2.6.0
python-crontab==3.2.0
coverage==7.6.0
django-cors-headers==4.4.0
flake8==7.1.0
debugpy==1.5.1
gunicorn==22.0.0
pillow==10.3.0
django-extensions==3.2.3
django-debug-toolbar==4.3.0
django-cachalot==2.6.3
django-redis==5.4.0